geoimport: import osm, move getting the heights in separte module
